mimxrt/flash: Splits mimxrt_flash.c.
Splits `mimxrt_flash.c` into low-level driver functions and MicroPython
code.
